Routes working with Moto Taxi and Hapi 17 new model
