+++ src/config/plugins/auth/index.js
