+++ src/config/plugins/mototaxi/index.js
@@ -2,9 +2,9 @@
 
 import * as mototaxi from 'mototaxi';
 import { AwilixResolutionError } from 'awilix';
-import container from '../container';
+import container from '../../container';
 
-import { commandHandlers } from '../../domain';
+import { commandHandlers } from '../../../domain';
 
 const plugin = {
   register: async (server: Object, options: Object) => {
