+++ src/config/registrations/mototaxi.js
@@ -1,0 +1,7 @@
+/* eslint-disable import/no-dynamic-require, global-require */
+export default [
+  {
+    // $FlowFixMe
+    plugin: require(`${__dirname}/../plugins/mototaxi/`).default,
+  },
+];
