--- src/config/registrationsT/mototaxi.js
@@ -1,5 +1,0 @@
-export default [
-  {
-    plugin: './config/plugins/mototaxi',
-  },
-];
