+++ src/routes/encounters/package.json
@@ -1,0 +1,4 @@
+{
+    "name": "encounters",
+    "version": "2.0.0"
+}
